Curs_8/01_Agent_langchain/requirements.txt
langchain
langchain-groq
langchain-community
python-dotenv
